backend/agent/tools/browser_demo.py
"""
Browser demonstration tool for testing browser-based interactions.
This tool allows the agent to open and interact with web content.
"""

from typing import Dict, Any, List, Optional
from agentpress.tool import Tool, SchemaType, ToolSchema
from utils.logger import logger

class BrowserDemoToolWrapper(Tool):
    """Tool for demonstrating browser-based interactions."""
    
    def get_schemas(self) -> Dict[str, List[ToolSchema]]:
        """Get the schemas for the tool."""
        return {
            "open_browser_demo": [
                ToolSchema(
                    schema_type=SchemaType.OPENAPI,
                    schema={
                        "type": "function",
                        "function": {
                            "name": "open_browser_demo",
                            "description": "Open a browser to demonstrate web-based interactions",
                            "parameters": {
                                "type": "object",
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "description": "URL to open in the browser"
                                    },
                                    "task": {
                                        "type": "string",
                                        "description": "Task to perform in the browser"
                                    }
                                },
                                "required": ["url"]
                            }
                        }
                    }
                )
            ]
        }
    
    async def open_browser_demo(self, url: str, task: Optional[str] = None) -> Dict[str, Any]:
        """
        Open a browser to demonstrate web-based interactions.
        
        Args:
            url: URL to open in the browser
            task: Optional task to perform in the browser
        
        Returns:
            Dictionary with result of the operation
        """
        try:
            logger.info(f"Opening browser demo with URL: {url}")
            # In a real implementation, this would launch a browser
            # For demo purposes, we'll just return success
            
            return {
                "success": True,
                "message": f"Browser opened successfully to {url}",
                "task": task or "No specific task provided",
                "browser_url": url
            }
        except Exception as e:
            logger.error(f"Error opening browser demo: {e}")
            return {
                "success": False,
                "message": f"Failed to open browser: {str(e)}"
            }
